spec: add centos-release, drop %{centos_rel} refs
I might be missing part of the picture here, but unlike its downstream
equivalent, there's no point in using a `%{centos_rel}` macro since
AFAICT this package is only used by rdgo. This was causing
`%{centos_rel}` to literally be in the various release files.

We also mimic the downstream version by adding a `/etc/centos-release`
and symlinking `redhat-release` and `system-release` to that.