spec: Bump fictional upstream release version to 7.5.0.1 (#8)
The `initscripts` package has:

```
Conflicts: redhat-release < 7.5-0.11
```

So this causes treecompose failures[1] because we're not providing a
high enough version. Let's fix this for now by just bumping the upstream
release version. I added the `.0.1` to make sure we definitely sort
higher than `7.5` regardless of the `Release` tag, which gets generated
by rdgo.

[1] https://github.com/CentOS/sig-atomic-buildscripts/issues/328